Add configuration file parsing and sanity checks
